Moved types closer to the code that uses them
