+++ src/provider.ts
@@ -1,5 +1,5 @@
-import { ClientRequestOptions } from "./types";
 import SDK from "./sdk";
+import { ClientRequestOptions } from "./client";
 
 export class ContextDataProvider {
 	getContextData(sdk: SDK, requestOptions?: Partial<ClientRequestOptions>) {
